Simplify Cursor definition
