+++ kernel-ewasm/cap9-core/src/lib.rs
@@ -32,49 +32,80 @@ pub trait Read<T> {
     fn read(&mut self, buf: &mut [T]) -> Result<(), Error>;
 }
 
-/// Reader that saves the last position.
-pub struct Cursor<T> {
-    inner: T,
-    pos: usize,
+
+// Seek does not seem to be implemented in core, so we'll reimplement what we
+// need.
+#[derive(Debug)]
+pub struct Cursor<'a, T> {
+    pub current_offset: usize,
+    pub body: &'a [T],
 }
 
-impl<T> Cursor<T> {
-    pub fn new(inner: T) -> Cursor<T> {
+impl<'a, T> Cursor<'a, T> {
+
+    pub fn new(body: &'a [T]) -> Cursor<'a, T> {
         Cursor {
-            inner,
-            pos: 0,
+            body,
+            current_offset: 0,
         }
+    }
+
+    // Read the byte at the cusor, and increment the pointer by 1.
+    pub fn read_ref(&mut self) -> Option<&'a T> {
+        if self.current_offset < self.body.len() {
+            let val = &self.body[self.current_offset];
+            self.current_offset += 1;
+            Some(val)
+        } else {
+            None
+        }
+    }
+
+    pub fn read_ref_n(&mut self, n: usize) -> &'a [T] {
+        let val = &self.body[self.current_offset..(self.current_offset + n)];
+        self.current_offset += n;
+        val
+    }
+
+    pub fn skip(&mut self, n: usize) {
+        self.current_offset += n;
+    }
+
+    pub fn remaining(&self) -> usize {
+        self.body.len() - self.current_offset
+    }
+
+    pub fn len(&self) -> usize {
+        self.body.len()
     }
 
     pub fn position(&self) -> usize {
-        self.pos
+        self.current_offset
+    }
+
+    pub fn inner(&self) -> &'a [T] {
+        self.body
     }
 }
 
-impl<T: AsRef<[u8]>> Read<u8> for Cursor<T> {
-    fn read(&mut self, buf: &mut [u8]) -> Result<(), Error> {
-        let slice = self.inner.as_ref();
-        let remainder = slice.len() - self.pos;
-        let requested = buf.len();
-        if requested > remainder {
+/// Implement standard read definition (which clones). This is basically the
+/// rust definition of read for slice.
+impl<'a, T: Copy> Read<T> for Cursor<'a, T> {
+    fn read(&mut self, buf: &mut [T]) -> Result<(), Error> {
+        if self.remaining() < buf.len() {
             return Err(Error::UnexpectedEof);
         }
-        buf.copy_from_slice(&slice[self.pos..(self.pos + requested)]);
-        self.pos += requested;
-        Ok(())
-    }
-}
+        let actual_self = &self.body[self.current_offset..];
+        let amt = core::cmp::min(buf.len(), actual_self.len());
+        let (a, _) = actual_self.split_at(amt);
 
-impl<T: AsRef<[U256]>> Read<U256> for Cursor<T> {
-    fn read(&mut self, buf: &mut [U256]) -> Result<(), Error> {
-        let slice = self.inner.as_ref();
-        let remainder = slice.len() - self.pos;
-        let requested = buf.len();
-        if requested > remainder {
-            return Err(Error::UnexpectedEof);
+        if amt == 1 {
+            buf[0] = a[0];
+        } else {
+            buf[..amt].copy_from_slice(a);
         }
-        buf.copy_from_slice(&slice[self.pos..(self.pos + requested)]);
-        self.pos += requested;
+
+        self.current_offset += amt;
         Ok(())
     }
 }
