+++ kernel-ewasm/cap9-kernel/src/lib.rs
@@ -221,7 +221,8 @@ pub fn call() {
             // call.
 
             // Put the input into a cursor for deserialization.
-            let mut input = Cursor::new(pwasm_ethereum::input());
+            let input_slice = pwasm_ethereum::input();
+            let mut input = Cursor::new(input_slice.as_slice());
             // Attempt to deserialize the input into a syscall. Panic on
             // deserialization failure.
             let syscall: SysCall = SysCall::deserialize(&mut input).unwrap();
